refactor: 회원 가입 uri 변경
