+++ src/main/java/com/becareful/becarefulserver/global/common/domain/BaseEntity.java
